requirements.txt
-e ./sphinx
sphinx-intl[transifex]>=0.9.11
